# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "HealthController_healthz",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        },
        "tags": ["health"]
      }
    },
    "/me": {
      "get": {
        "operationId": "MeController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["me"]
      }
    },
    "/engagements": {
      "get": {
        "operationId": "EngagementsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EngagementResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["engagements"]
      }
    },
    "/engagements/{id}/audit": {
      "get": {
        "operationId": "EngagementsController_audit",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditPageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["engagements"]
      }
    },
    "/auth/callback": {
      "get": {
        "operationId": "AuthController_callback",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallbackResponse"
                }
              }
            }
          }
        },
        "tags": ["auth"]
      }
    }
  },
  "info": {
    "title": "FDE Context Platform API",
    "description": "HTTP seam — WorkOS auth + request-scoped tenant/engagement context.",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "the opaque session token from /auth/callback"
      }
    },
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          }
        },
        "required": ["status"]
      },
      "MeResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["active", "disabled"]
          }
        },
        "required": ["userId", "tenantId", "email", "name", "status"]
      },
      "EngagementResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "endCustomerName": {
            "type": "string"
          },
          "regionPin": {
            "type": "string",
            "enum": ["us", "eu"]
          },
          "retentionPolicy": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["active", "closed", "shredded"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "endCustomerName", "regionPin", "retentionPolicy", "status", "createdAt"]
      },
      "AuditRowResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "engagementId": {
            "type": "string",
            "nullable": true
          },
          "actorType": {
            "type": "string"
          },
          "actorId": {
            "type": "string",
            "nullable": true
          },
          "action": {
            "type": "string"
          },
          "resourceType": {
            "type": "string",
            "nullable": true
          },
          "resourceId": {
            "type": "string",
            "nullable": true
          },
          "authzDecision": {
            "type": "object",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "engagementId",
          "actorType",
          "actorId",
          "action",
          "resourceType",
          "resourceId",
          "authzDecision",
          "reason",
          "createdAt"
        ]
      },
      "AuditPageResponse": {
        "type": "object",
        "properties": {
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditRowResponse"
            }
          },
          "nextCursor": {
            "type": "string",
            "description": "pass back as ?cursor for the next page"
          }
        },
        "required": ["rows"]
      },
      "CallbackResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "opaque session token — also set as the fde_session cookie"
          },
          "userId": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": ["token", "userId", "tenantId", "email"]
      }
    }
  }
}